+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+ - RoutingCache fix for potential infinite recursion
+
 ## [14.2.0] - 2019-02-14
 ### Added
  - Added DM-TX-4KZ-202-C and DM-TX-4KZ-302-C transmitters